(RE-6866) Update the bin directory to have symlinks
Currently INSTALLDIR\bin directory has .lnk shortcut
files in it. These will not work to start puppet or
anything, so we need to update it to use hardlinks.

In order to get symlinks we had to add a customaction
for each link. The actions will just run:

      cmd.exe /c mklink /H "link" "target"

The c flag stops a window from opening and H specifies
it is a hard link to a file

The custom actions were added to the sequences after the
InstallFinalize step because these links should likely
be created at the end