(RE-6892) Move the .bat files location
Before this we were planning on keeping the .bat files
that actually execute puppet in windows inside Puppet\puppet\bin
and creating symlinks in Puppet/bin to the .bat files. However
we have since decided that the symlinks are not the way to proceed
since they require custom actions for both install and uninstall
and the act of symlinking is moving away from being windows native.

Thus we elected to move the .bat files entirely to the Puppet\bin
directory and change the contents to point to the correct paths.

This update does the following:

* Adds each .bat file we were using in Puppet-Agent to the actual
Puppet-agent vanagon project. this means those files are now under
the resources/files/windows and we will no longer use those generated
by any of the install.rb scripts.

    * AS A SPECIAL NOTE: this PR itself does not explicitly
    deprecate those .bat files created by the install.rb
    scripts. that work (to remove those files from Puppet/puppet/bin)
    will come as a different PR.

* Updates the .bat files to point to the correct location of all the
executables in Puppet/puppet/bin