(PA-279) Add logic for puppet-agent short service names
Currently, puppet-agent services on osx use a short name instead
of a fully qualified name (puppet instead of com.puppetlabs.puppet).
This breaks the vanagon logic that determines the service names, and
thus the service restarts on package upgrades.

As we can't change the service names until the next major, this
adds some logic to the OSX preinstall to correctly handle service
restarts on package upgrade.