(PA-186) components should use upstream and rewrite urls
Prior to this commit, for anybody not on the Puppet, Inc VPN a person
had to manually gather sources for ~20 items and place them at a
location web accessable and then modify component files to point to that
location just to get started on a build.

Vanagon has had http rewrite rule magic in it for a while. This commit
simply puts the upstream URL in place for primary url on most
components. This should remove the first hour or two of work from any
community member attempting to build out the puppet-agent.

If you're a community member, now you'll simply have to comment out the
rewrite line in configs/projects/puppet-agent.rb

This will not allow community members to build hauwei, or aix due to
other things required that our only on our build infrastructure.

Windows will not work with just this commit, as other changes are
required. That will be addressed in another PR.