(maint) Add java to f23 build
Although we're not really building server packages for fedora at the
moment, that is something we'd like to be more on top of. This commit
adds the java dependency to the facter build. This will build facter
with JNI support, which will allow puppetserver to access facter via
facter.jar