(PA-287) Preserve hiera.yaml on upgrade
Since the new packages don't own hiera.yaml in codedir we need to back
that up during the preinstall steps and move it back into place in the
codedir in the postinstall steps.

It is important to copy first then move to avoid errors on removing the
old version of puppet agent.