(PA-350) Build pxp-agent with all static libs on Windows
Our packaging on Windows ends up with duplicate binaries in multiple
places for Leatherman and libgcc/libstdc++. It also creates
libcpp-pcp-client as a DLL with no expectation we'll ever use it apart
from pxp-agent.

Since we're not sharing any binary symbols, it would be simpler and
smaller to make pxp-agent a monolithic binary. Change to build that on
Windows.