(CPR-354) Add debian-8-armhf platform definition
This is a fairly normal platform definition. Keys are that it is a
cross-compile, we build on amd64. To enable that, you also need binutils
for the cross-compiled-target on the build architecture. In this case
that means installing binutils-arm-linux-gnueabihf as part of the
setup for use during dh_strip calls of packaging.