(PA-375) Disable locales on SLES and EL-4
Our images of these platforms come with an old version of gettext
installed. When the build attempts to use it, we run into errors because
expected gettext features are not available. Disable using gettext and
locales on these platforms.

This should have no other notable effects, as we use the default locale
on non-Windows platforms.