(PA-392) Ensure linking against PL libs
Prior to this commit, there wasn't any verification that we were linking
against our  libstdc++ vs the system-wide library. We'd like to make
sure that we've either statically linked libstdc++ and libgcc_s into our
builds or that we're using the runtime-provided components that would
end up in /opt/puppetlabs/puppet/lib.