(PA-396) Runtime should be more dynamic
Currently the runtime component of puppet-agent is very hard-coded and
not very flexible. It assumes very specific sonames and versions (all of
which match GCC 4.8.2). Since we now have several targets using GCC
5.2.0 and we've elected to enable the runtime component on all Linux
platforms, we need something more flexible.

This commit adds in a runtime.sh script that runs on the target to
determine the proper libraries and links to create in the puppet agent
libdir. It also enables runtime on all linux platforms.