(RE-7593) Update MCO to fix windows issues
There are two issues currently with the MSI build on MCO: First, the
mcollectived daemon is being installed to puppet/bin, which is wrong on
windows. this commit updates the mco config to ship to mcollective/bin on
windows platforms.

Second, openssl installs dlls to the puppet/bin, but when the mcollective
service is running there is no way to get that path to the ruby openssl
executables. So we need to copy those executables to sys/ruby/bin while we are
still installing ruby and puppet in different directories.