(RE-7667) OSX: Patch 'include/memory' for libc++ bug
Prior to this commit, cpp-hocon builds fail on OS X due to a known issue
(https://llvm.org/bugs/show_bug.cgi?id=18843) in an earlier version of
the libc++ headers shipped with LLVM (in the Xcode Command Line Tools).

This issue is fixed in Xcode Command Line Tools 7.3. Unfortunately,
installation of Xcode CLT 7.3 is blocked on OS X 10.9 and 10.10.

As a workaround, this commit updates puppet-agent's OS X 10.9 and 10.10
platform configurations to patch libc++'s 'include/memory' header as a
provisioning step before the build.