(maint) Reset puppet to a pre-corrective-change-reporting SHA
This resets the Puppet SHA to a version before corrective change
landed. This will let us get a clean build to the PE integration team
while we sort out what we're going to do in corrective change land.

This can't be our final shipping SHA since it's missing some other
fixes, but it should be good enough to unblock the integration team.