(PA-466) Correct env_windows_installdir
In commit 24f8e0d we introduced a different path for
FACTER_env_windows_installdir - it used to "C:\\Program Files\\Puppet
Labs\\Puppet" but now goes to "C:\\Program Files\\Puppet Labs\\Puppet\\facter".
This commit fixes this. The impact is that in PE, the aio_agent_build fact uses
this fact to create the path to the VERSION file - without it, no
aio_agent_build fact is reported, causing CI pipelines to fail.