(PA-466) correct env_windows_installdir
A previous commit incorrectly set the env_windows_installdir fact to be
suffixed with 'puppet'. This was tested with the aio_agent_build fact available
in PE, but it turns out that is actually incorrect and the combination of the
two returned a false positive. This commit updates the path to the correct
(pre-vanagon) path.