(PA-496) Update Registry entries to write to the correct place
The Wix files that read and write registry entries during upgrade to remember
preferences were writing the remembered properties to the wrong place
(HKLM\SOFTWARE\Puppet Labs\Puppet). This commit updates them to write to the
correct key location (HKLM\SOFTWARE\Puppet Labs\PuppetInstaller).

This commit also adds extra registry searches for the incorrect location
of the keys so we can still pull preferences from the incorrect locations
in the 1.6.2 release when upgrading from 1.6.1 to 1.6.2