(PA-495) Set rpath for CMake projects
Previously facter and pxp-agent set CMAKE_INSTALL_RPATH to ensure rpath
is set when installing to non-default locations. Packaging relies on
this to get the correct rpath configuration. Leatherman was missing this
configuration, which was exposed when moving to dynamically linking
against `libstdc++`.

Fix link paths by explicitly setting RPATH for all CMake projects
(in conjunction with settings in the toolchain fails, in accordance with
https://cmake.org/Wiki/CMake_RPATH_handling's "Always full RPATH"
scenario).