(PA-538) Add COMMITTERS,CONTRIBUTING from Puppet
This commit adds the COMMITTERS.md and CONTRIBUTING.md files from the
github.com/puppetlabs/puppet project as a baseline for ones to use for Puppet
Agent. This commit adds the files as-is without any modification from it's form
in Puppet, so that we can more easily validate changes that are specific to
Puppet Agent in one git diff.

Signed-off-by: Moses Mendoza <moses@puppet.com>