(PA-538) Modify COMMITTERS.md for Puppet Agent
This commit modifies the COMMITTERS.md file copied from
github.com/puppetlabs/puppet for Puppet Agent. Specifically, references to
Puppet are replaced with Puppet Agent. Additionally, notice of a 'security'
branch is removed, because Puppet  Agent does not have a security branch nor is
this part of the current workflow. We also remove reference to YARD docs as the
Puppet Agent project does not have these.

Signed-off-by: Moses Mendoza <moses@puppet.com>