(maint) Store elevate.exe locally instead of downloading it from buildsources
elevate.exe is used so that the Windows start menu items can run puppet
in an elevated context, in the same user context that the background
service runs as.

We're including it locally instead of downloading it from buildsources to
unblock CI while the http source checksum verification in vanagon is
refactored.