(PA-633) Copy "ensure_version_file" test from puppet
This commit adds the "ensure_version_file.rb" test from puppet/acceptance/tests
to puppet-agent. This test is an example of a test that belongs in puppet-agent
rather than puppet - it's not testing puppet functionality but rather
functionality that puppet-agent is supposed to provide (installing a version
file in the right place).

Signed-off-by: Moses Mendoza <moses@puppet.com>