(MAINT) Use Homebrew env. vars to prevent 💩
Homebrew 1.1.0 removes SHA1 sums (fine, whatever), but it also removes
the ability to run `brew` as root. Yes, this is a good idea but we
have to refactor the entire macOS build process to fix that from our
end.

But "we don't upgrade Homebrew! We use a snapshot!" you say. Well,
true. But `brew` will automatically, and silently, upgrade itself in
the background when you tap a new Homebrew keg... unless you tell it
not to.

While I was in there, I also turned off emoji and turned up verbosity.
Might as well, right?

Yes, this is specific to macOS Sierra right now. It's a snowflake.