(PA-247) Don't halt pxp-agent child processes
Disables halting pxp-agent child processes when restarting or stopping
the service. This makes behavior consistent with all other platforms for
pxp-agent.

Pickup pxp-agent fix from PCP-276 required for this to work.

Avoiding TerminateProcess opens the possibility for bugs in pxp-agent
that prevent it from shutting down when service halt is requested. A
deadlock can now occur that causes the pxp-agent to be unresponsive to
Windows SCM.