(MAINT) We should support CPPFLAGs and CFLAGS
A lot of our upstream dependencies expect things like include path
declarations (-I/some/path) to be in CPPFLAGS, not CFLAGS. The
expectation is that CFLAGS is reserved for compiler or
language runtime configuration like optimization (-Ox) . We work
around this a lot of the time by including CFLAGS inside of CXXFLAGS,
but just using CPPFLAGS lets them both inherit whatever is in CPPFLAGS
by default.

This commit specifically includes CPPFLAGS inside CFLAGS for backwards
compatibility while we work out where to set the appropriate values in
our dependency chain.