(PA-768) Make puppetres.dll permanent
puppetres.dll is managed and installed using the default file harvest
function, so make a separate puppetres.wxs file to define this as a
permanent component along with its registry entry.

The puppetres.dll file also needs to be filtered out of the harvest
operation. This is achieved by making a local copy of the
filter.xslt.erb file and adding the a filter statement for puppetres.dll

A future implemntation of vanagon could add a "filter_file" method which
would allow the filtering to be stated in puppet.rb and the specific
filter.xslt.erb file removed.