(PA-437) Add wrapper scripts to clean up linker environment
Previously, we relied on symlinks from our public bin directory to the
internal one to indicate which executables were public. This worked for
its purpose, but could cause a user's linker environment variables to
affect our executables. Because we rely on our own versions of many
libraries, this can cause hard-to-debug failures for our customers.

We now use a small wrapper script that cleans up the linker environment
variables, instead of a symlink. This maintains our clean separation
between public and internal executables, and makes sure that users
global settings won't mess with Puppet.

Users that want to adjust Puppet's behaviors with linker variables can
directly call the internal path to the executables. This does not
require us to stabilize those internal locations, as any user
intentionally messing with our loaded libraries has already broken out
of being on a supported configuration.