(PA-1023) Add step to generate MO files for puppet component
The gettext tooling for translation of user-facing strings in Puppet
looks for binary MO files containing translated strings for user's
locale. This commit adds a step to the configure step of the puppet
component to take PO files with translations and generate the binary
equivalents, placing them in puppet/locales/<locale>/LC_MESSAGES, one
for each locale. These files are then installed to the proper package
location along with the rest of Puppet's files in a later step.