(FACT-1624) Update Leatherman and Facter to include Ruby 2.4 fixes
In order to support Ruby 2.4, leatherman removed references to Bignum
and Fixnum (see LTH-124) in favor of Integer, which necessitated
changes in Facter. Since this was an API break, these updates need
to be brought into PA together for validation, so this commit updates
both components simultaneously, instead of via the separate component
pipelines.