(PA-1168) Patch Stomp to detect failed connections and reconnect
Previously failed connections would take minutes or hours to detect,
then leave MCO in an unusable state. Patch Stomp to handle failed
connections as soon as the timeout limits are reached and reconnect
cleanly.

The specific problem is an issue with Ruby/OpenSSL interaction, where
it's not thread-safe to close a socket while reading from it (which is
precisely what the Stomp gem does). A simple reproduction of that issue
can be found at https://gist.github.com/richardc/b5a1d4d28c8bbae24cd5.