(PA-1169) Backup and restore an existing environment hiera.conf
If there is an existing environment hiera.conf, we want to backup/
restore it, and also use it as an indication that the user has a
hiera config already and we should not lay down our new hiera
settings.

This also switches from `cp` to `mv` which will preserve permissions.