(maint) Update mirrored nssm to match upstream archive
Previously, our mirrored copy of nssm was a slightly altered archive
containing a VERSION file and missing a directory level, compared to the
upstream artifact. Since we no longer need the VERSION file for
packaging, this commit updates the md5 sum to match the upstream
version, which we will now mirror internally as-is.