(PA-1580) Enable testing of system openssl linked puppet agent on rhel7 platforms
The install script in all relevant components including puppet-agent repo
needs to be updated so it upgrades the version of openssl package on rhel7 platforms.
RHEL7 images used by CI/VMpooler uses an old openssl version, compared to that on
Centos7 which is where PA is built on,  which causes agent setup and tests to fail.
This has been preventing PA built against openssl to be tested in CI pipelines
which require one test target to be at least RHEL7.

Following components have been updated with above changes that puppet-agent now
refers to:
  - puppet-agent
  - puppet
  - hiera
  - facter
  - pxp-agent

The openssl package version to upgrade to can be specified using the following
environment variable: RHEL7_OPENSSL_VERSION. If not specified it defaults to
a version that is known to work with PA built against system openssl (on centos7).