(PA-1264) Add support to build PA against system openssl on FIPS platforms
This is being done to allow using puppet agent on FIPS enabled systems.
Vendored openssl is not allowed on FIPS systems as it is not FIPS certified.
System openssl linking is done when building on redhat_fips-7-x86_64 platform.
The rhel7 fips vmpooler image used for building will be FIPS enabled and
have the appropriate version of openssl-devel package installed on it.

Also adding redhat-fips-7-x86_64 as a build platform.