(PA-1264) Enable puppet agent to build against system openssl
This is being done to allow Puppet use in FIPS environments. Linking
to system ssl is based on an environment variable USE_SYSTEM_OPENSSL.
This feature will only be supported on el6, el7, f24 and f25 platforms.
For other platforms, the flag would be ignored and the build would continue
with vendor-ed openssl.