(PA-1617) Enable testing PA on systems running in FIPS mode
PA build against system openssl (PA-1264) need to be tested on platform
running in FIPS mode. This is a series of changes first requiring the above
and then upgrading rhel7 hosts to use newer version of openssl (PA-1580).
Enabling FIPS mode on target platforms is the final step in the process.
It could be followed by adjusting any tests that are incompatible with FIPS
like ones using MD5 which would fail.

This change currently only makes the agent host to run in FIPS mode when
running puppet acceptance tests.