(PA-1699) Use gcc 4.x to compile sqlite3 gem on Solaris
Previously, the vendored ruby tests failed on i386 and SPARC for Solaris 10,
and only i386 for Solaris 11 on the puppet-agent master pipelines. This
was due to the fact that the 5.3.x and master pipelines ship Ruby 2.4.2
with the agent. Unfortunately gcc 3.4.x, used on the failing Solaris
machines, does not work with Ruby 2.4.2. However gcc 4.x, used on
Solaris 11 SPARC (the passing case), does. This commit uses gcc 4.x for
all the Solaris platforms.