(PA-1402) Avoid loading custom facts during upgrade
Prior to this commit, we saved the output of
`puppet resource service puppet` and then applied that output with
`puppet apply`.

After this commit, we only save the the state of running or stopped
and then apply that with `puppet resource service`.  This avoids
`puppet apply` which can load custom facts that use the package
manager and can cause the upgrade to deadlock.