(maint) Add VERSION file for next release
The automation we are building for releases requires each repository
to know what version it is /going to be/. This was previously not
possible in puppet-agent, since it relied exclusively on git tags to
determine its version at build time.

All the build bits still use tags. The VERSION file is used only for
the release and tagging automation at this time.