(maint) Update to use packaging as a gem
Rather than cloning packaging to ext/packaging, we should be using
packaging as a gem. This allows us to not directly depend on
artifactory, as that is only a dependency of packaging and also vastly
simplifies logic in the Rakefile. Support for the PACKAGING_VERSION
environment variable has also been added to support runtime switching of
which packaging we're building/shipping/etc with.