(PA-1245) Add defs for cfacter-source-gem project
This will be a new vanagon project included with the agent. It makes use of 4
components:

* cpp-hocon-source < just the entire cpp-hocon source copied in to place
* leatherman-source < just the entire leatherman project copied in to place
* facter-source < just the entire facter project copied in to place
* cfacter-source-gem < a component with build rules to construct the source gem
from sources and resource files.

This commit is entirely additive, and should not effect the agent in any way.

The cfacter source gem is a gem that can be installed on a system with new
enough gcc/boost/cmake to give the user access to facter.

This commit makes use of changes from
https://github.com/puppetlabs/vanagon/pull/543