(PA-1894) Add custom vendor dir for shared rubygems
Puppet has a number of gem dependencies that are shared between using
it in a server or client context. These dependencies are currently
duplicated in both the agent and server packages, since both need to
be able to load Puppet's ruby code.

The problem with this is that adding new dependencies to Puppet
requires a dance with server releases, in order to ensure new versions
of the puppet code can still run with older servers that do not
include the gems, and in order to ensure that the latest server always
ships the gems needed to enable all of Puppet's features.

Instead, we are introducing a new vendored gem directory which will
contain gems intended to be used on both the agent and server. This
will allow us to ship the required gems along with the shared Puppet
code, ensuring the necessary gems for new features will always be
present.

There are a couple of special requirements here that mean we don't
want to use the normal vendor_ruby/gems dir:

1. This path needs to be ruby-version-independent, since it needs to
   be a stable API between agent and server packages even across major
   versions.
2. Because this path will be loaded by both MRI and JRuby, it needs to
   contain only pure-ruby gems. We want to keep the standard
   vendor_ruby/gems path available to ourselves for MRI-specific gems.