(PA-1896) Enable builds for Raspian/Debian 9 on armhf
This commit enables builds for Debian 9 armhf, which is the target for
running raspian 9 (raspberry PI Debian). This build also has the
advantage of not requiring anything from pl-build-tools as this compiles
all natively using the xcc toolchain provided by Debian.

Since this uses system yaml-cpp and boost, it does add runtime
requirements on those libraries (as well as libicu).