(PA-1918) Remove mcollective
We will no longer be shipping mcollective in puppet-agent 6.

The main reason for the deprecation is that many of mcollective's use
cases (triggering puppet runs, running remote tasks) are captured by
Orchestrator and Puppet Tasks. Thus, we are removing it from the
puppet-agent package so that it can be released separately from the
Puppet Platform.