(PA-1982) Calculate facter gem version on build host
Previously, the facter gem version was being computed in Vanagon. This
required fetching the facter version file from github, which was
leading to us hitting API rate limits.

This commit changes the build to calculate the version on the build
machine, using a small Ruby script. It avoids the need to hit github
and still creates artifacts with the proper version.