(PA-1995) Load puppet-runtime settings from yaml
Updates the puppet-agent.json file to a new format, which contains a
location (an http or file URL to some puppet-runtime output directory)
and a version for puppet-runtime. This information is used to determine
a download location for:

- an appropriate runtime tarball, and
- a yaml file containing inheritable project settings for the current platform

This removes the need to git-clone the runtime repository to inherit
settings on every build.