(maint) Modules should be installed from gh repo
This commit alters how we're handling modules during the install
process. Prior, we had been installing them from the forge. This commit
changes the install so that we're installing from a github repo. This
allows us to include development versions of modules in the puppet-agent
package, as we do for all other component projects included here.

This commit also modifies mailalias and maillist to read data from
corresponding json files.