(PA-2019) Restrict permission to Administrators and System
Restrict permissions to most of the subtree within
%ProgramData%\PuppetLabs so that LocalSystem and the local
Administrators group have full control. All other users are implicitly
denied. The resulting permissions are generally:

    BUILTIN\Administrators:(OI)(CI)(F)
    NT AUTHORITY\SYSTEM:(OI)(CI)(F)

The OI and CI mean object inherit and container inherit so that child
files and directories, respectively.

The following exceptions are made to allow Everyone read and execute:

 * %ProgramData%\PuppetLabs: The directory itself allows Everyone read
   and execute because the Client Tools MSI creates a configuration file
   within client-tools, and it must be readable by non-privileged users.

 * %ProgramData%\PuppetLabs\facter: Non-privileged users often configure
   facter to load facts from the privileged location, though that is not
   the default behavior.

The resulting permissions are:

    Everyone:(OI)(CI)(RX)
    BUILTIN\Administrators:(OI)(CI)(F)
    NT AUTHORITY\SYSTEM:(OI)(CI)(F)

During an initial install, Windows Installer will create the directories
and set permissions using the SDDL strings we specify. The resulting
DACL for the main PuppetLabs directory will be protected so that
inherited ACEs from %ProgramData% don't grant additional permissions to
children, e.g. %ProgramData%\PuppetLabs\code, beyond what we specify.

During an upgrade, the DACL for directories we are explicitly managing
will be protected, and any inherited ACEs will be removed from those
directories only, but not its children!! Permissions are only
automatically removed from child objects, when they are removed from the
parent where inheritance starts[1].

And due to the way SeChangeNotifyPrivilege works[2], users do not need
execute permission to traverse ancestor directories between
%ProgramData% and the %ProgramData%\PuppetLabs\path\to\something!!

As a result of both things, users can still access files within a
protected directory. To remedy this, use a custom action to reset
permissions on the child directories and files below each protected
directory so that inherited ACEs are recomputed based on the now
protected directory.

This commit switches from the WiX Permissions element to PermissionsEx
so that we can specify the desired permissions using an SDDL string,
which is not sensitive to different locales.

[1] https://msdn.microsoft.com/en-us/library/windows/desktop/aa376326(v=vs.85).aspx
[2] https://blogs.technet.microsoft.com/markrussinovich/2005/10/19/the-bypass-traverse-checking-or-is-it-the-change-notify-privilege/