(PA-2019) Use deferred custom action
Previously, the icacls custom action was executed before the files were
installed, so install would fail on a new machine, and the permissions
would be clobbed when applying the SDDL. Instead use a WIX quiet
deferred custom action to run icacls after files are installed and SDDL
permissions applied. See [1] for more info about immediate and deferred
execution.

The quiet action is necessary so that msiexec does not display a console
window when executing icacls.exe.

The sequence "execute" is necessary so that icacls doesn't run twice
during immediate and deferred phases.

Deferred actions do not have access to most MSI properties, so the
`SetProperty` is needed during the immediate phase to record the command
line that the `WixQuietExec` custom action will need later.

We "check" the return code from the custom action so that if icacls
fails, we abort the install.

Impersonation is "no" because we need to run in an elevated context to
reset permissions.

This also adds a custom action to reset permissions within facter's data
directory, as there could be files left over.

[1] http://www.installsite.org/pages/en/isnews/200108/index.htm