(PA-2019) Ensure System has permission on puppet's datadir
During an upgrade, the custom action to reset permission runs as System,
which due to PUP-6729, may not have permission to portions of the
agent's datadir, resulting in an access denied error.

This commit adds two custom actions to take ownership of puppet's
datadir, which grants the caller the WRITE_DAC permission to write the
DACL. It then runs icacls to grant system and local admins group full
control on all file/children.

This commit also chains the "After" properties so that the actions fire
in a predictable sequence.