(PA-2019) Assign ownership to the Administrators group
Assign the owner of files/directories to the local Administrators group
instead of the user that the process is running as. This is necessary
because the custom action runs as System, and results in files that are
not accessible to foreground runs of puppet, e.g. `puppet agent -t`.

Since System is an implicit member of Administrators, puppet will
continue working when running as a system or as administrator.