(PA-2019) Skip permissions reset during removal
Skip takeown/grant/reset logic if the package is being removed using the REMOVE
property[1]. Both the SetProperty and its corresponding CustomAction are skipped
together, though they need to be configured in different wxs files.

This affects uninstalls and upgrades as well since we always perform a major
upgrade, which means the currently installed MSI is removed, and a new one is
installed. The MSI being removed will skip the reset logic, but the MSI being
installed will perform the logic.

[1] https://msdn.microsoft.com/en-us/library/windows/desktop/aa371194(v=vs.85).aspx