(PA-2019) Be lenient about resetting permissions
Previously, running the agent as a service and later interactively through
cygwin could cause future attempts to restore permission inheritance to
fail due to permissions within the clientbucket directory.

When puppet runs as a service, it creates the clientbucket directory
with permissions:

    clientbucket BUILTIN\Administrators:(F)
                 NT AUTHORITY\SYSTEM:(RX)
                 Everyone:(Rc,S,RA)
                 CREATOR OWNER:(CI)(IO)(F)
                 CREATOR GROUP:(CI)(IO)(RX)
                 CREATOR OWNER:(OI)(IO)(R,W,D,WDAC,WO,DC)
                 CREATOR GROUP:(OI)(IO)(R)

Note *only* the "CREATOR OWNER/GROUP" ACEs are inherited by its children.
If you later run puppet interactively using cygwin and backup a file
into the local filebucket using:

    puppet resource file <path> ensure=absent force=true

Then the directory within the clientbucket will have permissions:

    clientbucket\5 X8IA34XBZ622S2I\Administrator:(I)(F)
                   CREATOR OWNER:(I)(CI)(IO)(F)
                   X8IA34XBZ622S2I\None:(I)(RX)
                   CREATOR GROUP:(I)(CI)(IO)(RX)
                   CREATOR OWNER:(I)(OI)(IO)(R,W,D,WDAC,WO,DC)
                   CREATOR GROUP:(I)(OI)(IO)(R)

Note *only* the local Administrator (owner) has access, but the local
Administrators group and LocalSystem do not. Cygwin is to blame because
the process token's Owner is set to Administrator instead of
Administrators. This behavior does not occur when running puppet
commands via cmd or powershell.

Commit 30849d04 attempts to restore permissions, but the custom action
running as LocalSystem does not have the "list directory" permission to
"5", causing an access denied error.

This commit makes the following changes:

* Don't try to take ownership of directories for which we don't have the
  list folder permission (/D N).

* Continue if an error occurs (/C)

* Display each processed file, instead of being quiet (/Q)

If the MSI fails to restore permissions, then the file will be left in
the same state it was before, so there isn't any loss in security, e.g.

    clientbucket\5\...\contents X8IA34XBZ622S2I\Administrator:(I)(F)
                                X8IA34XBZ622S2I\None:(I)(R)